Fix should_print pytest comparision
So, I'm not super happy with this, but I think it is the best tradeoff.

Before this patch, FakeStringIO.__eq__ was .strip'ing inconsently in
relation to the pytest_assertrepr_compare supporting it. So a couple
things would happen.
1) print("hello world") wouldn't match "hello world" or "hello world\n"
2) because the __eq__ implementation was inconsistent, the test would fail
   but show you an error saying "hello world\n" == "hello world\n" was false.

So this just strips everything all the time...

This is more sloppy than I would like to be, but this is just a starting
point to help the students get off the ground with testing things that
would require mocking print. I would rather it be too flexible with
whitespace on the outsides of the strings and make testing
their code easier.