Add --num param to what_have_i_done
I want the fish_greeting for the students to be a quick refresher
on their recent commands. So, I thought what_have_i_done was
almost what I needed. I just needed to add a way to limit how far
back it would look.

So this is a larger commit than I would like, but it ended up
hitting a couple things.

1) Looking at ~/.bash_history was outdated, so I changed it to
   calling out to `fish -c "history"` instead. At first I thought
   this was because of the shell buffering writes to the file, but
   it might also have to do with `fish` probably not writing to
   ~/.bash_history =). Either way, fish will be installed on the
   machines and I'm happy enough with `fish -c "history"`

2) Some refactoring to keep my sanity when trying to add the
   --num parameter.

3) Add click as a dependency for cli argument parsing.